[chan_lcr] Fixed broken communication with chan_lcr
There are two modes:

The remote mode is used to route a call directly to Asterisk or route a
call from Asterisk to an ISDN interface. (the old way)

The interface mode is used to handle chan_lcr as an interface, so calls
from Asterisk can be routed and behave like they come from an ISDN
interface. (documentation for that will follow)