Handle MT_L2IDLE
MT_L2IDLE is send by the mISDNuser library when all L2 processes
are gone. We could sent a MT_L2RELEASE if we are the L2 master and
do not want permanent L2 monitoring. This would deactive L2 and also
shutdown L1.

For now we do nothing, only log a debug message.

Signed-off-by: Karsten Keil <kkeil@linux-pingi.de>